requirements.txt
beautifulsoup4==4.9.3
Pillow==7.0.0
urllib3==1.25.8
